debugging - Still not working
Pulled all files that relate to mininet out of the virtual box so I can follow the functions to the places they were defined.

Also started from scratch with the script, because I wasn't finding the cause of the error it was having. But it's still not working for a reason or another...

To make the current script I simply used miniedit to make the base template of the devices, and I was planning on changing the position and mobility of the nodes as needed, as well as adding controller commands to test the network. However nothing worked, even within miniedit.

PS: The mininet files are too big for me to commit, but they're all the mininet specific folders from the VM they provide on the official github.